Optimizations:
- Remove normalization on output. This was entirely unnecessary.
- Optimize `with_lightness` to reuse arrays.
- Only do HSL conversion if necessary.

| Configuration    | Before | After | Note                        |
| ---------------- | ------ | ----- | --------------------------- |
| H=0 S=0 L=0      | 150ms  | 0     | Pass through is free now    |
| H=0 S=0 L=-35    | 710ms  | 90ms  |                             |
| H=0 S=0 L=-33    | 1000ms | 140ms |                             |
| H=0 S=29 L=0     | 820ms  | 510ms |                             |
| H=0 S=-39 L=0    | 740ms  | 480ms | Saved a `np.clip` over S=29 |
| H=-62 S=24 L=-18 | 1000ms | 830ms |                             |

Tested on a 8468x4320 RGB image.

<details>

Before:
![image](https://github.com/chaiNNer-org/chaiNNer/assets/20878432/1a095180-69b7-4c7c-ab19-5d9b5bff42a5)


After:
![image](https://github.com/chaiNNer-org/chaiNNer/assets/20878432/e60befc0-602b-45b1-9bd2-4db201870821)

</details>